backend/BackDjango/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.2

DateTime~=4.3
pip~=20.2.3
toml~=0.10.2
configparser~=5.0.2
wheel~=0.37.0
cryptography~=3.4.7
lxml~=4.6.3
tornado~=6.0.4
keyring~=23.0.1
colorama~=0.4.4
Jinja2~=2.10.3
numpy~=1.21.2
setuptools~=49.2.1
coreapi~=2.3.3
uritemplate~=3.0.1
coreschema~=0.0.4
requests~=2.23.0
Markdown~=3.2.1
Pygments~=2.7.1
pandas~=1.3.2
sklearn~=0.0
scikit-learn~=0.23.2
PyMySQL~=0.9.3